backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
pydantic-settings
email-validator
aiosqlite
starlette
google-auth
google-auth-oauthlib
requests
requests-aws4auth
